autotrader/requirements.txt
alpaca-py
yfinance
pandas
pandas-ta
finnhub-python
transformers
torch
xgboost
scikit-learn
backtrader
quantstats
streamlit
plotly
apscheduler
python-dotenv
SQLAlchemy
psycopg2-binary
bcrypt
joblib
